PDFBOX-5660: Sonar - Fix sign-extension issues in bitmap bit manipulation by using int for bitwise operations

+++ src/main/java/org/apache/pdfbox/jbig2/decoder/mmr/MMRDecompressor.java
@@ -572,7 +572,7 @@ public class MMRDecompressor
         {
 
             final int offset = currentOffsets[index];
-            byte value;
+            int value;
 
             if ((index & 1) == 0)
             {
